adds support for namespaced operations
So far, the client would only operate in one namespace, the hardcoded `admin` namespace. With this commit, you can now explicitly specify in which namespace to operate. The API is backwards compatible in that, if you don’t specify a namespace on a dot (and implicitly on a branch) then it defaults to the `admin` namespace.